requirements.txt
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
Flask==3.0.3
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
psycopg==3.2.3
psycopg2-binary==2.9.10
typing_extensions==4.12.2
Werkzeug==3.1.1
requests ==2.32.3
Flask-Mail==0.10.0
